Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-pannernode-interface/panner-automation-equalpower-stereo-expected.txt

CONSOLE MESSAGE: line 37: TypeError: undefined is not an object (evaluating 'panner.positionX.value = x')

Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'panner.positionX.value = x')

PASS # AUDIT TASK RUNNER STARTED. 
PASS > [test] Equal-power panner model of AudioPannerNode with stereo source 
